# perf: cache view methods