These are nicely cacheable, and don't have to be part of the class.

```

codegen         mean runtime:  221.53ms, runs:   255.83,  215.28,  197.83,  201.14,  219.83,  205.83,  206.70,  220.71,  198.30,  293.85
methodcache     mean runtime:  165.73ms, runs:   210.85,  159.71,  155.94,  158.69,  169.23,  158.60,  165.78,  150.29,  175.99,  152.22
profile         mean runtime:  916.79ms, runs:   805.64,  902.98,  891.13,  902.14,  938.34,  939.29,  997.60,  923.44,  923.32,  944.03

to

codegen         mean runtime:  200.05ms, runs:   270.67,  186.91,  177.84,  172.52,  183.54,  196.58,  184.18,  249.55,  179.50,  199.21
methodcache     mean runtime:  160.23ms, runs:   151.52,  160.38,  148.77,  151.17,  149.59,  150.05,  165.42,  156.59,  202.07,  166.77
profile         mean runtime:  731.61ms, runs:   698.33,  721.21,  721.92,  691.91,  721.54,  821.83,  780.41,  736.11,  709.44,  713.36
```

No idea why the diff is shot, I de-indented and added a cache decorator.