@tinyb0t 2023-09-04
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/shape/shapetracker.py            30              27
------------------------------------------------------------
lines added in the tinygrad folder: 3
```