@geohot 2023-09-04
Is this more or less readable?