@wozeparrot 2023-09-04
This is fine, the github diff makes it seem really bad. If you check in codespaces or in vscode it renders the proper diff.