@roelofvandijk 2023-09-04
![image](https://github.com/tinygrad/tinygrad/assets/3604013/2b3c3ed6-2697-40ee-bc2c-e4b86b36c0ce)