@roelofvandijk 2023-09-04
There's a slightly faster version of this, where I pass only the actual parts of the view that are used,to each function. Better cache hits but that's a bit less readable.