@geohot 2023-09-04
Less readable